requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.23.0
pandas>=1.5.0
pillow>=9.0.0
opencv-python>=4.7.0
imageio[pyav]>=2.30.0
scikit-image>=0.19.0
scipy>=1.11.0
torchmetrics>=1.0.0
lpips>=0.1.4
einops>=0.6.0
tqdm>=4.65.0
